# improve the performance of reverse nucleotide sequence