Before (master):
```
julia> seq = randdnaseq(100_000)
100000nt DNA Sequence:
TGAAGACATAACACGTGAGTGGGCAGGTGACACAGCACCGAGAG…AACCATCTGGACGGTAGACACGAGACATTAACCACTACTCTCTT

julia> @benchmark reverse($seq)
BenchmarkTools.Trial:
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  48.94 kb
  allocs estimate:  3
  minimum time:     34.84 μs (0.00% GC)
  median time:      37.50 μs (0.00% GC)
  mean time:        47.93 μs (9.23% GC)
  maximum time:     2.99 ms (95.88% GC)

julia> seq = randdnaseq(99_999)
99999nt DNA Sequence:
CACCCTCAGAAAGATGACCGAGTAGCGCAGTCGAAAAGAGTAGA…GGAACTAGTCAGCGATGCTAACGGTGCACTCTGTCAATCATTTG

julia> @benchmark reverse($seq)
BenchmarkTools.Trial:
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  48.94 kb
  allocs estimate:  3
  minimum time:     34.80 μs (0.00% GC)
  median time:      37.34 μs (0.00% GC)
  mean time:        47.65 μs (10.09% GC)
  maximum time:     3.46 ms (98.87% GC)

```

After:
```
julia> seq = randdnaseq(100_000)
100000nt DNA Sequence:
AGAAATTCTTGAGCGCAAGGTGTGGTCGGCGCGACTGGTTTGGC…CGACCTGTCTCCACCTGAGTCGGAGTCTGCAGTCCTCTAAGGTC

julia> @benchmark reverse($seq)
BenchmarkTools.Trial:
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  48.94 kb
  allocs estimate:  3
  minimum time:     18.68 μs (0.00% GC)
  median time:      21.23 μs (0.00% GC)
  mean time:        29.74 μs (15.56% GC)
  maximum time:     3.05 ms (98.83% GC)

julia> seq = randdnaseq(99_999)
99999nt DNA Sequence:
TTCCCCAATGGTACATAGCAAATGCAAGAAAGGAGTCCCTGACG…CGGCAAGTATCTAGTAGGGAAGCAAGACTATGATGCAAGACGTG

julia> @benchmark reverse($seq)
BenchmarkTools.Trial:
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  48.94 kb
  allocs estimate:  3
  minimum time:     27.53 μs (0.00% GC)
  median time:      30.31 μs (0.00% GC)
  mean time:        39.22 μs (11.25% GC)
  maximum time:     2.90 ms (98.19% GC)

```